Don't crash of when radius proxy is not config.
